Refactor code - Split GCSFileInputPlugin class to PluginTask, GcsFileInput and SingleFileProvider class
